Missed long to int in AisTarget

+++ ais-communication/src/main/java/dk/dma/ais/data/AisTarget.java
@@ -122,7 +122,7 @@ public abstract class AisTarget implements Serializable {
         country = Country.getCountryForMmsi(aisMessage.getUserId());
     }
 
-    public long getMmsi() {
+    public int getMmsi() {
         return mmsi;
     }
